map default repositories to api classes
Previously: `defaultRepositories` were wrapped in a private `ApiRepo` class.
Now: `defaultRepositories` are mapped to their api equivalences.
reason:
It seems that ApiRepo is not really used for anything else and there is a method that can core.Repository to api.Repository. Without this change we cannot extract any information from `defaultRepositories` using this interface.